requirements.txt
pyttsx3
pyautogui
ddgs
beautifulsoup4
requests
SpeechRecognition
pyaudio
pyperclip
opencv-python
python-pptx
edge-tts==1.2.2